requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
sentence-transformers
faiss-cpu
pandas